sample-apps/11-client/api/requirements.txt
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
Werkzeug==2.2.2
zipp==3.8.1
mysql-connector==2.2.9
opentelemetry-exporter-otlp==1.9.1
opentelemetry-instrumentation-flask==0.28b1
opentelemetry-instrumentation-mysql==0.28b1
opentelemetry-instrumentation-requests==0.28b1
opentelemetry-instrumentation-logging==0.28b1
opentelemetry-sdk==1.9.1
protobuf==3.15.6
